Fix broadcasts, by refreshing the API IDs
The `get_api_by_name` method changes it's results depending on whether
the WS is logged-in or not (any call to `.login` with empty creds will
do).

This commit changes `send` so when a `login` call resolves, the API IDs
get refreshed.

This closes #31 and probably is the cause of #37.